Added more running time classes.

+++ src/main/java/fi/helsinki/coderodde/msc/RunningTime.java
@@ -3,5 +3,5 @@ package fi.helsinki.coderodde.msc;
 interface RunningTime {
     
     double estimate(final double entropy,
-                    final double fingers);
+                    final int fingers);
 }
